# account-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Account Microservice API
  description: API for managing accounts
  version: 1.0.0
tags:
  - name: Account
paths:
  /api/v1/accounts:
    get:
      tags:
        - Account
      summary: Get all accounts
      operationId: getAllAccounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'

    post:
      tags:
        - Account
      summary: Create a account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/movements:
    get:
      tags:
        - Movement
      summary: Get all movements
      operationId: getAllMovements
      responses:
        '200':
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movement'

    post:
      tags:
        - Movement
      summary: Create a movement
      operationId: createMovement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequest'
      responses:
        '201':
          description: Movement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/accounts/{id}:

    get:
      tags:
        - Account
      summary: Get account by ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Account
      summary: Update account
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Account
      summary: Delete account
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Account deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/movements/{id}:

    get:
      tags:
        - Movement
      summary: Get movement by ID
      operationId: getMovementById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Movement found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Movement
      summary: Update movement
      operationId: updateMovement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequest'
      responses:
        '200':
          description: Movement updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movement'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Movement
      summary: Delete movement
      operationId: deleteMovement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Movement deleted
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          example: 1
        accountNumber:
          type: string
          example: 123456
        accountType:
          type: string
          enum:
            - Savings
            - Checking
          example: Savings
        initialBalance:
          type: number
          example: 100.00
        customerId:
          type: integer
          example: 12345
        status:
          type: boolean
          example: true

    Movement:
      type: object
      properties:
        id:
          type: integer
          example: 1
        accountId:
          type: integer
          example: 1
        movementDate:
          type: string
          format: date-time
          example: "2024-05-20T14:30:00Z"
        movementType:
          type: string
          enum:
            - Deposit
            - Withdrawal
          example: Deposit
        amount:
          type: number
          example: 100.00
        balance:
          type: number
          example: 100.00
        

    AccountRequest:
      type: object
      required:
        - accountNumber
        - accountType
        - customerId
      properties:
        accountNumber:
          type: string
          example: 123456
        accountType:
          type: string
          enum:
            - Savings
            - Checking
          example: Savings
        initialBalance:
          type: number
          example: 100.00
        customerId:
          type: integer
          example: 12345
        status:
          type: boolean
          example: true
    MovementRequest:
      type: object
      required:
        - accountId
        - movementDate
        - movementType
        - amount
      properties:
        id:
          type: integer
          example: 1
        accountId:
          type: integer
          example: 1
        movementDate:
          type: string
          format: date-time
          example: "2024-05-20T14:30:00Z"
        movementType:
          type: string
          enum:
            - Deposit
            - Withdrawal
          example: Deposit
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          example: 100.00

    # Esquema de error estandar
    ErrorDetails:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetails'